fix: allows partial copying of password fields (#6426)
Previously copying behavior was changed to copy select full fields
https://github.com/bitwarden/clients/pull/5709#discussion_r1247033019

This was done to prevent additional spaces from being added to
the start or end of passwords but this prevented partial selection.

This commit implements partial selection. By wrapping generated
username and password fields in `<div>` tags instead of
`<span>` tags additional text nodes will not be part of the
selection.